fix(DESCRIPTION): include license type
Previously the default value from usethis:: was in the License portion
of the DESCRIPTION file. And, thanks to @k-doering-NOAA for making me
aware that we had a license file. `check()` is no longer giving me
errors.
Close #47
If anyone wishes to propose that we change the type based on those
that are listed in the above issue then we should open a new issue
and everyone would need to agree on the new type. See guidance at
https://r-pkgs.org/license.html